Refact: api-query rq, rs interface 분리
