+++ src/interfaces/auth/IPostSignUp.ts
@@ -1,0 +1,11 @@
+export interface IPostSignUpRequest {
+  name: string;
+  email: string;
+  password: string;
+}
+
+export interface IPostSignUpResponse {
+  name: string;
+  email: string;
+  hashedPassword: string;
+}
