+++ tsconfig.json
@@ -25,6 +25,7 @@
       "@public/*": ["./src/public/*"],
       "@components/*": ["./src/components/*"],
       "@types/*": ["./src/types/*"],
+      "@interfaces/*": ["./src/interfaces/*"],
       "@styles/*": ["./src/styles/*"],
       "@utils/*": ["./src/utils/*"],
       "@hooks/*": ["./src/hooks/*"],
